backend/requirements.txt
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary
protobuf
python-dotenv
pydantic
pydantic-settings
pytest
pytest-cov
